+++ iPERCore/tools/human_pose2d_estimators/openpose/dataset.py
@@ -7,7 +7,7 @@ import os
 
 
 def normalize(img, img_mean, img_scale):
-    img = np.array(img, dtype=np.float32)
+    img = np.array(img, dtype=float)
     img = (img - img_mean) * img_scale
     return img
